requirements.txt
Flask==3.0.0
Jinja2==3.1.2
Click==8.1.7
PyYAML==6.0.1
pydantic==2.5.0
gunicorn==21.2.0
pytest==7.4.3
flake8==6.1.0
